Fix daylight savings. Assume UTC.
When we add a file to the database, we currently parse the string as if it's in
the current time zone. Then we save this parsed timestamp's Unix time. For
nonexistent times like 2015-03-08 02:30:00 GMT-0800, we save the UTC Unix time
for 2015-03-08 03:30:00 GMT-0800. So multiple data points are saved for this
timestamp, because the actual 3:30am datum is also parsed as 3:30am.

The chart retrieves this Unix time from the database and formats it as the
current time zone, then displays it. So the Unix time remains constant
throughout this whole process, it just matters how we parse the initial string.