Make sidebar scrollable. Polish the scrolling experience.
The experience of scrolling the sidebar is now identical to the experience of
scrolling the charts. The changes:

- You can now scroll the sidebar. It scrolls independently of the chart area.
- The scrollbar is positioned correctly. I had to work around a Chromium bug.
- The scrollbars are the same color. Previously the scrollbar in the charts area
  was white because the operating system chose that color based on the
  background color. The background color is a gradient, which causes the OS to
  be weird and choose white (even though it chooses black if you use either
  gradient endpoint as a solid-colored background). Now that scrollbar lives in
  a child element, this no longer happens. This is all a little bit goofy.

Additionally, the app no longer bounces up and down when you scroll to the top
and bottom (like a webpage). Now it feels more like a real desktop app, similar
to Atom. The bounces also had the issue of showing weird rendering artifacts
sometimes.