Never translate a timestamp into another time zone
In this change, I mimic Python's "naive" datetimes by removing time zone info
from timestamp strings, then parsing it as UTC. The UI will always render the
timestamp as provided, never switching its time zone. UTC is the new null.

Similarly, on export the results will be the same timestamps that were provided,
but the time zone will not be included. Previously the time zone was always
converted to UTC, shifting the timestamps.

Also, I stopped using moment.js in places where we're not parsing or formatting
strings. It's easier to understand whether the inputs are a string/Date/number
when we're not calling the magical moment() function.