Control our own y scale
Now every time the chart draws, it checks back to see what y minimum and maximum
it should use.

I chose a policy of: use a fixed scale with the actual maximum and minimum, but
if the data isn't occupying at least 20% of the vertical space, adjust the scale
so that we're using 20% of the vertical space. Also, make room for green anomaly
bars. The data will no longer crash into the green anomaly bars.

It will be easy to change any of this policy now that it's all hooked up.

This involved changing ModelData.jsx so that more information is available to
the yScaleCalculate callback.

I made the plugin code a little clearer -- we just import it, we don't need to
get the Dygraph function from it.