Fix server defaults for DATETIME and TIMESTAMP columns in taurus.engine, because 0 ("0000-00-00 00:00:00") is deprecated as default for those colum types as of mysql 5.7.8, and will fail with mysql installed with default config
Fix server defaults for DATETIME and TIMESTAMP columns in htm.it, because
0 ("0000-00-00 00:00:00") is deprecated as default for those colum types
as of mysql 5.7.8, and will fail with mysql installed with default config.

Point tests in app/runtime/aggregator_metric_collection_test.py that require running instances to
the us-west-2 region, since we have no more running test EC2 instances in us-east-1

Fix default in instance_status_history.timestamp column of the htmengine's sleleton-app database.

Switch to fabric==1.11.1 (from 1.9.0) to avoid pulling in unstalbe paramiko==2.0

Update configure-htm-it-pipeline-centos7-ami and install-htm-it-from-bare-centos7.sh to use uwsgi==2.0.12 instead of 2.0.4 for consistency with package requirements.