Always use model result timestamps in navbar
Before this change, the experience of running a model was different
based on whether the model was aggregated. The time navigation bar
would sometimes stay constant with the selection moving across it, and
other times would add timestamps of model results to the time
navigation bar as they arrived. Now we always do the latter.