New concept: "resolution". No more jittering.
This removes the "displayPointCount", similar to how we replaced
"pointsPerBar" with buckets. Rather than specifying a number of points
to the chart, specify a timespan and the meaning of that timespan.
e.g. "5 minutes per anomaly bar" or "5 years per chart width".

Similar to before, this is needed when the data isn't uniformly
distributed (e.g. if it contains gaps).

Also fix a bug in computeGapThreshold: array.sort() sorts by string!
e.g. [1, 3, 2, 10].sort() // [1, 10, 2, 3]