On export, use original time zone
This change adds more fields to the database, but it does not grow the
store. These ISO timestamps live in the database, and they're only
used during export. The MetricData timestamp is never used, but this
keeps things consistent, and we have the option of using it later.

For running models, this change moves the timestamp parsing
upstream. Before sending a timestamp to the client, node will compute
the "naive time". So we're no longer sending a timestamp string
through IPC.